refactor: greatly improve the package types
This is actually a breaking change in terms of typescript, but since I only released version 1.0 yesterday I am going to make an exception and do a major bump instead.
